pyproject.toml
[project]
name = "fai"
version = "0.1.0"
description = "Real-time AI character video interaction"
requires-python = ">=3.12"
dependencies = [
    "openai",
    "python-dotenv",
    "sounddevice",
    "opencv-python",
    "numpy",
    "anthropic>=0.77.1",
    "elevenlabs>=2.34.0",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pre-commit",
    "ruff",
    "mypy",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_ignores = true

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
addopts = "--cov=fai --cov-report=term-missing --cov-fail-under=80"

[tool.coverage.run]
branch = true
source = ["fai"]
omit = ["fai/__main__.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if __name__ == .__main__.:",
    "raise NotImplementedError",
]
fail_under = 80
show_missing = true

[dependency-groups]
dev = [
    "pytest-cov>=7.0.0",
]
